Implemented WsMsg::status(), backed by status.rs
