+++ backend/src/ws/mod.rs
@@ -1,3 +1,4 @@
 mod ws_actor;
 pub mod ws_conn;
 pub mod ws_message;
+mod ws_structs;
